Handle customer default rich menu cases

+++ app/interactions/rich_menus/delete.rb
@@ -5,11 +5,17 @@ module RichMenus
     object :social_rich_menu
 
     def execute
-      response = LineClient.delete_rich_menu(social_rich_menu)
-
-      if response.is_a?(Net::HTTPOK) || response.is_a?(Net::HTTPNotFound)
-        social_rich_menu.image.purge_later if social_rich_menu.image.attached?
+      if Rails.env.development?
         social_rich_menu.destroy
+      else
+        response = LineClient.delete_rich_menu(social_rich_menu)
+
+        if response.is_a?(Net::HTTPOK) || response.is_a?(Net::HTTPNotFound)
+          social_rich_menu.image.purge_later if social_rich_menu.image.attached?
+          social_rich_menu.destroy
+        else
+          errors.add(:social_rich_menu, :delete_failed)
+        end
       end
     end
   end
